flower_fraud_detection/pyproject.toml
[tool.poetry]
name = "flower-fraud-detection"
version = "0.1.0"
description = "Production-grade Federated Learning for Fraud Detection using Flower Framework"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
flwr = "^1.11.0"
torch = "^2.0.0"
numpy = "^1.24.0"
hydra-core = "^1.3.0"
omegaconf = "^2.3.0"
pandas = "^2.0.0"
scikit-learn = "^1.3.0"
tensorboard = "^2.13.0"
torch-tb-profiler = "^0.4.0"
tqdm = "^4.65.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
isort = "^5.12.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
